v1.3.16 - attempt to fix drm master race condition
When EmulationStation starts lr-mame with a selected game, dmarquees recieves the game rom name and immediately tries to reset the crtc for marquee display. This probably happens at the same time lr-mame is attempting to grab drm master and initialize it's own frame buffer(s). Unfirtunately it seems that lr-mame is unforgiving and does not retry - it simply exits with an error (of which I can't seem to prove or see in any logs after fighting with chatgpt) ... anyway, we're going to attempt to delay the reset crtc call and also monitor when lr-mame takes and releases master. After the lr-mame release or timeout, we call the reset crtc function (fingers crossed!).